help-komennon korjaaminen: - Siirretty testit omaan moduuliin. Lisätty testit, että komento triggeröityy oikein vain jos se on viestin alussa, sen edellä on aloitusmerkki ja sen jälkeen ei ole muuta - Toteutettu korjaus regexiin, että ei tapahdu jos yllä oleva ei toteudu
